services/commerce-service/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12

# Database
motor==3.5.0
pymongo==4.8.0

# Cache
redis[hiredis]==5.0.0

# Payments
stripe==8.0.0

# Authentication
PyJWT==2.9.0

# Configuration & Validation
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
mongomock-motor==0.0.34
fakeredis[json]==2.21.0
